refactor: domain, service 분리
