+++ src/main/java/lotto/domain/ErrorMessage.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.domain;
 
 public enum ErrorMessage {
     HEADER {
