+++ src/main/java/lotto/domain/LottoStatus.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.domain;
 
 public enum LottoStatus {
     START(1),
