+++ src/test/java/lotto/LottoTest.java
@@ -1,7 +1,7 @@
 package lotto;
 
-import lotto.model.ErrorMessage;
-import lotto.model.Lotto;
+import lotto.domain.ErrorMessage;
+import lotto.domain.Lotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
