+++ src/test/java/lotto/domain/WinningNumberTest.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
